requirements.txt
langchain
langchain-core
langchain-huggingface
langchain-openai
langchain-community
sentence-transformers
python-dotenv
faiss-cpu
youtube-transcript-api
faster-whisper
streamlit
google-search-results
pytesseract
transformers==4.41.2
huggingface-hub==0.34.6
fastapi
uvicorn
python-multipart
reportlab